refactor: Split CreateResourcesV2Handler from ResourcesResponderV2 (DEV-3341)
